Inserter: Prevent unnecessary re-renders when selecting a different block
